Add trait for XML documents and use it in Document class
